Refactor hss adapter cache into separate file, add class comment
